add flink udf hook
