+++ linkis-engineconn-plugins/flink/flink-core/src/main/java/org/apache/linkis/engineconnplugin/flink/client/sql/operation/OperationFactoryImpl.java
@@ -49,6 +49,7 @@ public class OperationFactoryImpl implements OperationFactory {
                 context, call.operands[0], Boolean.parseBoolean(call.operands[1]));
         break;
       case CREATE_TABLE:
+      case CREATE_FUNCTION:
       case DROP_TABLE:
       case ALTER_TABLE:
       case CREATE_CATALOG:
