+++ linkis-engineconn-plugins/flink/flink-core/src/main/java/org/apache/linkis/engineconnplugin/flink/client/sql/operation/impl/DDLOperation.java
@@ -59,6 +59,9 @@ public class DDLOperation implements NonJobOperation {
       case CREATE_TABLE:
         actionMsg = "create a table";
         break;
+      case CREATE_FUNCTION:
+        actionMsg = "create a function";
+        break;
       case CREATE_DATABASE:
         actionMsg = "create a database";
         break;
