+++ linkis-engineconn-plugins/flink/flink-core/src/main/java/org/apache/linkis/engineconnplugin/flink/client/sql/parser/SqlCommand.java
@@ -46,6 +46,8 @@ public enum SqlCommand {
 
   CREATE_DATABASE,
 
+  CREATE_FUNCTION,
+
   ALTER_DATABASE,
 
   DROP_DATABASE,
