+++ linkis-engineconn-plugins/flink/flink-core/src/main/java/org/apache/linkis/engineconnplugin/flink/client/sql/parser/SqlCommandParserImpl.java
@@ -152,6 +152,9 @@ public class SqlCommandParserImpl implements SqlCommandParser {
     } else if (node instanceof SqlCreateDatabase) {
       cmd = SqlCommand.CREATE_DATABASE;
       operands = new String[] {stmt};
+    } else if (node instanceof SqlCreateFunction) {
+      cmd = SqlCommand.CREATE_FUNCTION;
+      operands = new String[] {stmt};
     } else if (node instanceof SqlDropDatabase) {
       cmd = SqlCommand.DROP_DATABASE;
       operands = new String[] {stmt};
