+++ linkis-engineconn-plugins/flink/flink-core/src/main/resources/linkis-engineconn.properties
@@ -18,5 +18,5 @@ wds.linkis.server.version=v1
 wds.linkis.engineconn.debug.enable=true
 #wds.linkis.keytab.enable=true
 wds.linkis.engineconn.plugin.default.class=org.apache.linkis.engineconnplugin.flink.FlinkEngineConnPlugin
-wds.linkis.engine.connector.hooks=org.apache.linkis.engineconn.computation.executor.hook.ComputationEngineConnHook
+wds.linkis.engine.connector.hooks=org.apache.linkis.engineconn.computation.executor.hook.ComputationEngineConnHook,org.apache.linkis.engineconnplugin.flink.hook.FlinkJarUdfEngineHook
 wds.linkis.engineconn.executor.manager.class=org.apache.linkis.engineconnplugin.flink.executormanager.FlinkExecutorManager
